@robodoo 2024-12-16
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/190739.png)](https://mergebot.odoo.com/odoo/odoo/pull/190739)