@pivi-odoo 2024-12-16
Note:
In small dataset, when having around 1k variants, performance is roughly equivalent. This patch just scales better for a large number of variants, potentially unblocking a timeout.